// project/ws/app/src/lib/routes/app-toc/routes/app-toc-certification/models/certification.model.ts
export interface ICertificationMeta {
  isCompleted: boolean
  warningMessage: string
  eligibility: ICertificationEligibility
  qualified: boolean
  lastTakenDate: ICertificationDate
  slotFreezeDate: ICertificationDate
  booking: ICertificationBooking
  verification_request: ICertificationVerificationRequest
  budget_Request: ICertificationBudgetRequest
  qualified_score?: string
}

export interface ICertificationEligibility {
  eligible: boolean
  reason: ICertificationEligibilityReason
}

export interface ICertificationEligibilityReason {
  code: number
  message: string
  nextRegisterDate?: ICertificationDate
}

export interface ICertificationDate {
  day: number
  month: number
  year: number
  timeZone?: string
}

export interface ICertificationBooking {
  booking_type: string
  dc?: string
  testcenter?: string
  country?: string
  location?: string
  date: ICertificationDate
  slot: string
  slotno: number
  status: string
  key?: string
  icfdId?: number
}

export interface ICertificationVerificationRequest {
  exam_date: number
  result_type: string
  result: string
  verifierEmail: string
  document: {
    document_url: string
    document_name: string
  }[]
  status: string
  reject_reason: string
}

export interface ICertificationBudgetRequest {
  manager_id?: string
  amount: number
  currency: string
  status: string
  reject_reason: string
}

export interface ISubmitOrWithdrawRequest {
  exam_date: ICertificationDate | number
  result_type: string
  result: string
  fileName: string
  verifierEmail: string
}

export interface IAccLocation {
  dc: string
  testcenters: string[]
}

export interface ITestCenterSlotList {
  dc: string
  testcenter: string
  slotdata: {
    date: number // timestamp
    slots: ITestCenterSlot[]
  }[]
}

export interface ITestCenterSlot {
  slot_no: number
  slot_time: string
  seats_available: boolean
}
// ...
export interface ICertificationRequestItem {
  record_type: TCertificationRequestType
  user?: {
    name: string
    email: string
  }
  currently_with: {
    name: string
    email: string
  }
  raised_on?: ICertificationDate
  status?: string
  certification: string
  certification_name: string
  booking_type?: string
  country?: string
  location?: string
  date?: ICertificationDate
  slot?: string
  slotno?: number
  user_contact?: string
  icfdid?: number
  sino?: number
  ecdp_id?: number
  amount?: number
  currency?: string
  exam_date?: ICertificationDate
  document?: {
    document_url: string
  }[]
  document_url?: string
  resultType?: string
  result?: string
  upload_id?: number
  sampleCertificationUrls?: string[]
}
// ...
export interface ICertificationSendResponse {
  res_code: number
  res_message: string
}
// ...
export interface ICertificationUserPrivileges {
  canProctorAtDesk: boolean
  canVerifyResult: boolean
  canApproveBudgetRequest: boolean
  manager: string
}

export type TCertificationRequestType =
  | 'result_verification'
  | 'budget_approval'
  | 'proctor_approval'
  | 'acc'
  | 'all'